+++ pytest.ini
@@ -1,2 +1,5 @@
 [pytest]
 log_cli = True
+markers = 
+    real_camera: marks tests that require a real camera to be connected  # use like -m "real_camera"
+    emulated_camera: marks tests that instead will emulate a camera  # inherits from the real_camera tests; use like -m "emulated_camera and not real_camera"
